// src/entities/formulaire-creation-dto.ts
export class FormulaireCreationDto {

    idLogiciel1: number;
    nomLogiciel1: string;
    repertoireLogiciel1: string;

    idLogiciel2: number;
    nomLogiciel2: string;
    repertoireLogiciel2: string;

    idLogiciel3: number;
    nomLogiciel3: string;
    repertoireLogiciel3: string;

    idLogiciel4: number;
    nomLogiciel4: string;
    repertoireLogiciel4: string;

    idLogiciel5: number;
    nomLogiciel5: string;
    repertoireLogiciel5: string;
    
    idSiteweb1: number;
    urlSiteweb1: string;

    idSiteweb2: number;
    urlSiteweb2: string;

    idSiteweb3: number;
    urlSiteweb3: string;

    idSiteweb4: number;
    urlSiteweb4: string;

    idSiteweb5: number;
    urlSiteweb5: string;

}